refactor(workstation): import layout atoms from their owners
Follow up on Chloe-JY’s re-export cleanup and PR #1350. Preserve layout atom implementations and the existing public layout API.
