+++ src/ActionSystem/actions/chatPanelActions.zod.ts
@@ -9,10 +9,8 @@ import {
   modelPickerStyleAtom,
 } from "@src/store/ui/chatPanelAtom";
 import { stationModeAtom } from "@src/store/ui/simulatorAtom";
-import {
-  chatPanelPositionAtom,
-  workStationLayoutModePersistAtom,
-} from "@src/store/ui/workStationAtom";
+import { chatPanelPositionAtom } from "@src/store/ui/workStationLayout/chatPositionAtoms";
+import { workStationLayoutModePersistAtom } from "@src/store/ui/workStationLayout/splitLayoutAtoms";
 import { getInstrumentedStore } from "@src/util/core/state/instrumentedStore";
 
 const emptyParams = z.object({});
