+++ src/modules/WorkStation/CodeEditor/Panels/EditorMainPane/hooks/useSourceControlPaneActions.ts
@@ -16,7 +16,7 @@ import { useCallback, useMemo, useState } from "react";
 import { useRefreshSpin } from "@src/hooks/ui/useRefreshSpin";
 import type { QuickAction } from "@src/modules/WorkStation/shared";
 import type { SourceControlFilterMode } from "@src/modules/WorkStation/shared/SidebarModules";
-import { workStationPrimarySidebarCollapsedPersistAtom } from "@src/store/ui/workStationAtom";
+import { workStationPrimarySidebarCollapsedPersistAtom } from "@src/store/ui/workStationLayout/primarySidebarAtoms";
 import {
   type GitReviewNavigationSnapshot,
   gitReviewNavigationAtom,
