+++ src/modules/WorkStation/CodeEditor/Panels/EditorPrimarySidebar/hooks/useExplorerTabs.ts
@@ -11,7 +11,7 @@ import React, { useCallback, useMemo } from "react";
 import { useTranslation } from "react-i18next";
 
 import { HugeiconsIcon } from "@src/icons";
-import { workStationPrimarySidebarTabAtom } from "@src/store/ui/workStationAtom";
+import { workStationPrimarySidebarTabAtom } from "@src/store/ui/workStationLayout/primarySidebarAtoms";
 
 import { ICON_CONFIG, PANEL_CONSTANTS, TAB_LABELS, TAB_ORDER } from "../config";
 import type { ExplorerViewMode } from "../types";
