+++ src/modules/WorkStation/shared/TabBarTrailingControls.tsx
@@ -15,10 +15,8 @@ import { useTranslation } from "react-i18next";
 import { TabBarTrailingIconButton } from "@src/components/TabPill/TabBarTrailingIconButton";
 import { HEADER_ICON_SIZE } from "@src/config/workstation/tokens";
 import { HugeiconsIcon, PencilRulerIcon, SidebarBottomIcon } from "@src/icons";
-import {
-  workStationDevToolsCollapsedAtom,
-  workStationEditorSecondaryCollapsedAtom,
-} from "@src/store/ui/workStationAtom";
+import { workStationEditorSecondaryCollapsedAtom } from "@src/store/ui/workStationLayout/bottomPanelAtoms";
+import { workStationDevToolsCollapsedAtom } from "@src/store/ui/workStationLayout/devToolsCollapsedAtoms";
 import { activeStatusBarCallbacksAtom } from "@src/store/ui/workStationLayout/statusBarAtoms";
 
 export const TabBarBottomPanelToggle: React.FC = memo(() => {
